third_party/blink/web_tests/dom/mutation-event-tests/fast/dom/subtree-modified-attributes-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMSubtreeModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
DOMSubtreeModified should fire when attributes are added or removed, but not modified (see bug 81141)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS expected is true
PASS expected is true
PASS expected is true
PASS expected is true
PASS successfullyParsed is true

TEST COMPLETE
